Struct kademlia_routing_table::NodeInfo

pub struct NodeInfo<T, U> where U: Eq + Hash {
    pub public_id: T,
    pub connections: HashSet<U>,
    // some fields omitted
}

A routing table entry representing a node and the connections to that node.

Fields

public_id

The information identifying the node.

connections

The connections to the node, e. g. sockets or other kinds of connection handles.

Methods

impl<T, U> NodeInfo<T, U> where T: PartialEq + HasName + Debug, U: Eq + Hash

fn new<V>(public_id: T, connections: V) -> NodeInfo<T, U> where V: IntoIterator<Item=U>

Creates a new node entry with the given ID and connections.

fn name(&self) -> &XorName

Returns the XorName of the peer node.