Crate kademlia_routing_table

Kademlia Routing Table

Kademlia Routing Table implementation.

Structs

NodeInfo
Allows user defined data for routing table

RoutingTable
The RoutingTable class is used to maintain a list of contacts to which the node is connected.

Constants

GROUP_SIZE
Defines the size of close group

PARALLELISM
Defines the number of contacts which should be returned by the target_nodes function for a target which is outwith our close group and is not a contact in the table.

Traits

HasName
required trait for the info held on a node by routing_table

Functions

bucket_size
Bucket size

group_size
Defines the size of close group

optimal_table_size
optimal table size as usize

parallelism
parallelism as u64